requirements.txt
numpy
matplotlib
tensorflow
scikit-learn
streamlit
pandas
altair
Pillow